intel-hub/requirements.txt
feedparser>=6.0
jinja2>=3.1
beautifulsoup4>=4.12
pyyaml>=6.0
certifi>=2024.0
